backend/requirements.txt
# WATCHDOG Backend Dependencies
# Python 3.9+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Management
python-dotenv==1.0.0

# HTTP Client for OpenRouter API
httpx==0.25.2

# Async Support
aiofiles==23.2.1

# Risk Engine Dependencies (Member2)
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3

# Policy Engine Dependencies (Member4)
jsonschema==4.20.0

# Optional: Direct LLM SDKs (if not using OpenRouter)
# openai==1.3.0
# anthropic==0.7.0

# Development Tools (optional)
# pytest==7.4.3
# httpx==0.25.2
# black==23.12.0
# ruff==0.1.7
